RR-1791 - Move code that groups strengths so that it can be re-used by other routes
