--- server/routes/profile/profileStrengthsFunctions.ts
@@ -1,18 +1,0 @@
-import type { AlnScreenerList, AlnScreenerResponseDto, StrengthResponseDto, StrengthsList } from 'dto'
-import { Result } from '../../utils/result/result'
-import dateComparator from '../dateComparator'
-
-const getNonAlnActiveStrengths = (strengths: Result<StrengthsList>): Array<StrengthResponseDto> =>
-  strengths.getOrNull()?.strengths.filter(strength => strength.active && !strength.fromALNScreener) ?? []
-
-const getLatestAlnScreener = (screeners: Result<AlnScreenerList>): AlnScreenerResponseDto =>
-  screeners
-    .getOrNull()
-    .screeners.sort((left: AlnScreenerResponseDto, right: AlnScreenerResponseDto) =>
-      dateComparator(left.screenerDate, right.screenerDate, 'DESC'),
-    )[0]
-
-const getActiveStrengthsFromAlnScreener = (alnScreener: AlnScreenerResponseDto): Array<StrengthResponseDto> =>
-  alnScreener?.strengths.filter(strength => strength.active && strength.fromALNScreener) ?? []
-
-export { getNonAlnActiveStrengths, getLatestAlnScreener, getActiveStrengthsFromAlnScreener }
